chore(packs): move export q modal to packs root
